rename image enricher module
