+++ packages/ragbits-document-search/src/ragbits/document_search/ingestion/enrichers/__init__.py
@@ -1,4 +1,4 @@
 from .base import ElementEnricher
-from .images import ImageElementEnricher
+from .image import ImageElementEnricher
 
 __all__ = ["ElementEnricher", "ImageElementEnricher"]
