+++ packages/ragbits-document-search/src/ragbits/document_search/ingestion/enrichers/image.py
